% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivity-costs.R
\name{add_prod_costs}
\alias{add_prod_costs}
\title{Add productivity costs to a cost-effectiveness object}
\usage{
add_prod_costs(ce, prod_costs)
}
\arguments{
\item{ce}{An object of class "ce".}

\item{prod_costs}{An object of class "prod_costs".}
}
\value{
Unknown
}
\description{
Add productivity costs to a 
\href{https://hesim-dev.github.io/hesim/reference/ce.html}{cost-effectiveness object}
from the \href{https://hesim-dev.github.io/hesim/index.html}{hesim} pacakge.
}
\seealso{
See the example in the \href{https://innovationvalueinitiative.github.io/IVI-NSCLC/articles/tutorial.html}{tutorial}.
}
