% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adverse_events}
\alias{adverse_events}
\title{Adverse events}
\format{A data.table object with one row for each adverse event and with
columns:
\describe{
  \item{ae_name}{Name of adverse event.}
  \item{ae_abb}{Abbreviation for the adverse event.}
}}
\usage{
adverse_events
}
\description{
Dataset of adverse events included in the model.
}
\examples{
print(adverse_events)
}
\keyword{datasets}
