% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_dist}
\alias{age_dist}
\title{Patient age}
\format{A \code{data.table} with columns:
\describe{
  \item{age_cat}{Age category.}
  \item{age_lower}{Lower bound of age category.}
  \item{age_upper}{Upper bound of age category.}
  \item{age_mid}{Midpoint of age category.}
  \item{prop}{Proportion of patients in age category.}
}
Also contains the attributes:
\describe{
\item{mean}{Mean age.}
\item{sd}{Standard deviation of age.}
}}
\source{
{\url{https://seer.cancer.gov/csr/1975_2015/results_merged/topic_age_dist.pdf}}
}
\usage{
age_dist
}
\description{
Distribution of patient ages.
}
\examples{
print(age_dist)
attr(age_dist, "mean")
attr(age_dist, "sd")
}
\keyword{datasets}
