% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costmods.R
\name{annualized_tx_costs}
\alias{annualized_tx_costs}
\title{Annualized treatment costs}
\usage{
annualized_tx_costs(x = iviNSCLC::params_costs_tx)
}
\arguments{
\item{x}{An object in the same format as \code{\link{params_costs_tx}}.}
}
\value{
An object of class "annualized_tx_costs", which is a
 \code{data.table} of annualized treatment costs. Contains the columns:
\describe{
\item{agent}{See description from  the \code{dosage} element from \code{\link{params_costs_tx}}.}
\item{treatment}{See description from the \code{dosage} element from \code{\link{params_costs_tx}}.}
\item{dosage}{See description from the \code{dosage} element from \code{\link{params_costs_tx}}.}
\item{dose}{See description from the \code{dosage} element from \code{\link{params_costs_tx}}.}
\item{unit}{See description from the \code{dosage} element from \code{\link{params_costs_tx}}.}
\item{units_per_day}{See description from the \code{dosage} element from \code{\link{params_costs_tx}}.}
\item{duration_days}{See description from the \code{dosage} element from \code{\link{params_costs_tx}}.}
\item{discount_lower}{See description from the \code{discount} element from \code{\link{params_costs_tx}}.}
\item{discount_upper}{See description from the \code{discount} element from \code{\link{params_costs_tx}}.}
\item{acquisition_costs}{Annualized drug acquisition costs.}
\item{administration_costs}{Annualized drug administration costs.}
}
}
\description{
Compute annualized drug acquisition and administration costs.
}
\examples{
annualized_tx_costs(iviNSCLC::params_costs_tx)
}
\seealso{
\code{\link{params_costs_tx}}
}
