% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costmods.R
\name{create_costmods}
\alias{create_costmods}
\title{Create cost models}
\usage{
create_costmods(params, struct, patients)
}
\arguments{
\item{params}{A "sampled_params" object returned from \code{\link{sample_params}}.}

\item{struct}{A \code{\link{model_structure}} object.}

\item{patients}{A data table returned from \code{\link{create_patients}}.}
}
\value{
An object of class "StateVals" from the 
\href{https://innovationvalueinitiative.github.io/hesim/}{hesim} package.
}
\description{
Create a cost model for four cost categories: (i) treatment costs (i.e., drug
acquisition and administration costs), (ii) inpatient medical costs, 
(iii) outpatient medical costs, and (iv) costs due to adverse events.
}
\examples{
# Parameters
params <- sample_params(n = 2)

# Treatment sequences
txseq1 <- txseq(first = "erlotinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "gefitinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2)

# Patient population
pats <- create_patients(n = 2)

## Model structure
struct <- model_structure(txseqs, dist = "weibull")

## Cost models
costmods <- create_costmods(params, struct, pats)
}
