% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{create_states}
\alias{create_states}
\title{Create health states}
\usage{
create_states(object)
}
\arguments{
\item{object}{Object used to create health states.}
}
\value{
The form of the returned value depends on \code{object}.
}
\description{
Generic function to create data table of health states.
}
\seealso{
\code{\link{create_states.txseq_list}}
}
\keyword{internal}
