% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{create_states.txseq_list}
\alias{create_states.txseq_list}
\title{Create health states tables}
\usage{
\method{create_states}{txseq_list}(object)
}
\arguments{
\item{object}{A \code{\link{txseq_list}} object.}
}
\value{
\describe{
\item{state_id}{The state ID number.}
\item{state_name}{The state name,}
\item{state_name_long}{A long-form state name.}
}
}
\description{
Create a list of data tables describing the health states for the model
structure associated with each modeled treatment sequence.
}
\examples{
txseq1 <- txseq(first = "erlotinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "osimertinib",
                second = c("PBDC", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2) 
create_states(txseqs)
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2, start_line = "second") 
create_states(txseqs)
}
