% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_disease.R
\name{create_strategies}
\alias{create_strategies}
\title{Create treatment strategies object}
\usage{
create_strategies(object, line = c("1", "2", "2+"), mutation = c("neg",
  "pos"))
}
\arguments{
\item{object}{The treatment sequences of interest. Must be objects of class
\code{\link{txseq_list}}.}

\item{line}{The line of treatment.}

\item{mutation}{If the line of treatment is > 1, whether the patient has
a T790M mutation.}
}
\value{
A \code{\link{data.table}}.
}
\description{
Create a \code{\link{data.table}} containing information on treatment strategies used
at a particular treatment line. Of the same format as the \code{strategies}
element in \link[hesim]{hesim_data} in the \code{hesim} package.
}
