% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{create_trans_mat}
\alias{create_trans_mat}
\title{Create transition matrix}
\usage{
create_trans_mat(object)
}
\arguments{
\item{object}{A \code{\link{model_structure}} object.}
}
\value{
A transition matrix of the same format as in the \link[mstate]{mstate} 
package.
}
\description{
Create a transition matrix describing patient transitions between 
health states.
}
\examples{
struct <- model_structure(n_states = "four", start_line = "first")
create_trans_mat(struct)

struct <- model_structure(n_states = "three", start_line = "first")
create_trans_mat(struct)

struct <- model_structure(n_states = "three", start_line = "second")
create_trans_mat(struct)
}
