% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{create_trans_mats}
\alias{create_trans_mats}
\alias{create_trans_mats.txseq_list}
\title{Create line-specific transitions matrices}
\usage{
create_trans_mats(object)

\method{create_trans_mats}{txseq_list}(object)
}
\value{
An object of class "trans_mats", which is a list of matrices with length 
equal to the number of treatment lines.
}
\description{
Create transition matrices in the same format as the \link[mstate]{mstate} 
package for each treatment line.
}
\examples{
txseq1 <- txseq(first = "erlotinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "gefitinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab")) 
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2)
trans_mats <- create_trans_mats(txseqs)
class(trans_mats)
print(trans_mats)
}
