% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{create_trans_mats.txseq_list}
\alias{create_trans_mats.txseq_list}
\title{Create transition matrices}
\usage{
\method{create_trans_mats}{txseq_list}(object)
}
\arguments{
\item{object}{A \code{\link{txseq_list}} object.}
}
\value{
A list of transition matrices for each treatment sequence, each 
in the same format as in the \link[mstate]{mstate} package.
}
\description{
Create a list of transition matrices describing patient transitions between 
health states for the model associated with each modeled treatment sequence.
}
\examples{
txseq1 <- txseq(first = "erlotinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "osimertinib",
                second = c("PBDC", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2) 
create_trans_mats(txseqs)
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2, start_line = "second") 
create_trans_mats(txseqs)
}
