% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilmod.R
\name{create_utilmod}
\alias{create_utilmod}
\title{Create utility model}
\usage{
create_utilmod(params, struct, patients)
}
\arguments{
\item{params}{A "sampled_params" object returned from \code{\link{sample_params}}.}

\item{struct}{A \code{\link{model_structure}} object.}

\item{patients}{A data table returned from \code{\link{create_patients}}.}
}
\value{
An object of class "StateVals" from the 
\href{https://innovationvalueinitiative.github.io/hesim/}{hesim} package.
}
\description{
Create a model for health state utility given
}
\examples{
## Treatment sequences
txseq1 <- txseq(first = "erlotinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "gefitinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2)

# Patient population
pats <- create_patients(n = 2)

## Model structure
struct <- model_structure(txseqs, dist = "weibull")

## Utility model
utilmod <- create_utilmod(params, struct, pats)
}
