% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivi-nsclc.R
\docType{package}
\name{iviNSCLC}
\alias{iviNSCLC}
\alias{iviNSCLC-package}
\title{iviNSCLC: an R package for running the IVI-NSCLC model.}
\description{
iviNSCLC: an R package for running the IVI-NSCLC model.
}
