% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcda.R
\name{lpvf}
\alias{lpvf}
\title{Linear partial value function}
\usage{
lpvf(x, x_min, x_max, score_min = 0, score_max = 100)
}
\arguments{
\item{x}{Performance of the criterion on the original scale.}

\item{x_min}{Minimum value of criterion on original scale. If higher performance
is better, then the minimum value is the lowest possible value; otherwise, it
is the highest.}

\item{x_max}{Maximum value of criterion on original scale. If higher performance
is better, then the maximum value is the highest possible value; otherwise, it
is the lowest.}

\item{score_min}{Minimum value of score on common scale.}

\item{score_max}{Maximum value of score on common scale.}
}
\value{
The value of the score on the common scale.
}
\description{
Convert performance for criterion on original scale to a score on a common scale
using a linear partial value function.
}
