% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{model_structure}
\alias{model_structure}
\title{Model structure}
\usage{
model_structure(n_states = c("four", "three"), start_line = c("first",
  "second"))
}
\arguments{
\item{n_states}{Number of modeled health states.}

\item{start_line}{The starting line of treatmnet that is being modeled. When
modeling second line treatment, the first line must be specified
in order to characterize a treatment history.}
}
\value{
A list
}
\description{
Define the structure of the oncology model.
}
\examples{
model_structure()
model_structure(n_states = "three", start_line = "first")
model_structure(n_states = "three", start_line = "second")
}
