% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mstate_nma_pfs}
\alias{mstate_nma_pfs}
\title{Progression-free survival from network meta-analysis}
\format{A data.table with the following columns:
\describe{
  \item{model}{The statistical model.} 
  \item{tx_name}{Name of the treatment.}
  \item{month}{Trial month.}
  \item{mean}{Mean PFS from the posterior distribution.}
  \item{l95}{Lower limit of 95 percent credible interval from posterior distribution.}
  \item{u95}{Upper limit of the 95 percent credible interval from the posterior distribution.}
}}
\usage{
mstate_nma_pfs
}
\description{
Curves of progression-free survival (PFS)---the duration of time a patient is alive without disease 
progression---estimated using a Bayesian multi-state network meta-analyis (NMA).
}
\examples{
head(mstate_nma_pfs)
}
\keyword{datasets}
