% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params_ae_nma}
\alias{params_ae_nma}
\title{Adverse event NMA parameters}
\format{A list where each element is a matrix for a distinct adverse event.
Each column is a distinct treatment and each row is a random draw from
the posterior distribution. Each matrix contains a "lookup" attribute with 
the abbreviations used in the columns for adverse events.}
\usage{
params_ae_nma
}
\description{
The posterior distribution of the probability of an adverse event by
treatment. Based on separate models by adverse event with a binomial 
likelihood and a logit link.
}
\keyword{datasets}
