% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params_costs_ae}
\alias{params_costs_ae}
\title{Adverse event cost parameters}
\format{A data table with the following columns:
\describe{
\item{ae_name}{Name of the adverse event.}
\item{ae_abb}{Abbreviation for the adverse event.}
\item{mean}{Mean costs.}
\item{lower}{2.5\% quantile for mean costs. Currently assumed to be +20\%
of mean costs given lack of data for estimating standard errors.}
\item{upper}{97.5\% quantile for mean costs. Currently assumed to be -20\%
of mean costs given lack of data for estimating standard errors.}
\item{se}{Standard error of costs.}
\item{ref}{Reference for mean costs.}
}}
\usage{
params_costs_ae
}
\description{
Mean costs by adverse event.
}
\examples{
print(params_costs_ae)
}
\keyword{datasets}
