% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params_costs_inpt}
\alias{params_costs_inpt}
\title{Inpatient cost parameters}
\format{A data table with the following columns:
\describe{
\item{state_name}{Name of the health state.}
\item{mean}{Mean outpatient costs.}
\item{se}{Standard error of outpatient costs.}
\item{ref}{BibTeX reference for estimate.}
}}
\usage{
params_costs_inpt
}
\description{
Mean monthly inpatient costs by health state.
}
\examples{
print(params_costs_inpt)
}
\keyword{datasets}
