% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params_mstate}
\alias{params_mstate}
\title{Multi-state model paramters}
\format{A list containing parameter estimates from Weibull, Gompertz,
and 2nd order fractional polynomial survival distributions. Within each distribution,
parameters are contained in a list of lists with the outer list denoting a 
treatment line (1L, 2L, 2L+) and the inner list 
a \code{\link[hesim]{params_surv_list}} object from the \code{hesim} package.}
\usage{
params_mstate
}
\description{
The parameters of the individual-level continuous time state transition model
estimated using the multi-state network meta-analysis.
}
\keyword{datasets}
