% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params_mstate_nma}
\alias{params_mstate_nma}
\title{Multi-state network meta-analysis parameters}
\format{A list of \code{\link[hesim]{params_surv}} objects from the 
\href{https://innovationvalueinitiative.github.io/hesim/}{hesim} package
package. The list contains regression coefficient estimates from the Weibull,
Gompertz, and 2nd order fractional polynomial survival distributions. The 
parameter (e.g., scale and shape for the Weibull distribution) of each survival 
distribution are predicted as a function of treatments, health state transitions, and
treatment history (for 2L treatments).}
\usage{
params_mstate_nma
}
\description{
Posterior distributions of the regression coefficients of the continuous 
time state transition model estimated using the multi-state network meta-analysis.
}
\keyword{datasets}
