% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{pkg_env}
\alias{pkg_env}
\title{Package environment}
\format{An object of class \code{environment} of length 4.}
\usage{
pkg_env
}
\description{
An environment to store global variables for the package.
}
\keyword{datasets}
