% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_disease.R
\name{sim_disease}
\alias{sim_disease}
\title{Simulate disease progression}
\usage{
sim_disease(txseqs, params, n_patients)
}
\arguments{
\item{txseqs}{The treatment sequences of interest. Must be objects of class
\code{\link{txseq_list}}.}

\item{params}{The posterior distribution of the parameters of the disease model. Must be an object of class
\code{\link[hesim]{params_surv}} from the \code{hesim} package.}

\item{n_patients}{The number of patients to model.}
}
\value{
An object of class "indiv_ctstm_disprog". See the description for \code{\link[hesim]{IndivCtstmTrans}} 
from the \code{hesim} package.
}
\description{
Simulate disease progression for EGFR+ non-small cell lung cancer patients
given treatment sequences of interest.
}
