% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy}
\alias{tidy}
\title{Tidy data}
\usage{
tidy(object, ...)
}
\arguments{
\item{object}{An object to create tidy data from.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
A generic function for creating tidy data.
}
\seealso{
\code{\link{tidy.ae_probs}}
}
\keyword{internal}
