% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txseq.R
\name{tx_2L}
\alias{tx_2L}
\title{Second line treatment options}
\usage{
tx_2L(first)
}
\arguments{
\item{first}{A first line treatment.}
}
\value{
A list of two elements:
\describe{
\item{pos}{Possible treatments with a T790M mutation.}
\item{neg}{Possible treatments without a T790M mutation.}
}
}
\description{
Possible second line treatment for EGFR+ NSCLC patients conditional on the first
line treatment.
}
\examples{
first <- tx_1L()[2]
print(first)
tx_2L(first)
}
