% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txseq.R
\name{txseq_list}
\alias{txseq_list}
\title{A list of treatment sequences}
\usage{
txseq_list(...)
}
\arguments{
\item{...}{Objects to form a list.}
}
\value{
An object of class "txseq_list", which is a list of objects of class
\code{\link{txseq}}.
}
\description{
Create a list of objects of class "txseq".
}
\examples{
txseq1 <- txseq(first = "erlotinib",
               second = c("osimertinib", "PBDC"),
               second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "gefitinib",
               second = c("osimertinib", "PBDC"),
               second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab")) 
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2) 
class(txseqs)        
print(txseqs$seq1)
}
