% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{create_patients}
\alias{create_patients}
\title{Create patient data table}
\usage{
create_patients(n)
}
\arguments{
\item{n}{Number of patients to model.}
}
\description{
Create a data table of patients to model.
}
\examples{
create_patients(n = 10)
}
