% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{create_states}
\alias{create_states}
\alias{create_states.txseq_list}
\title{Create health states table}
\usage{
create_states(object)

\method{create_states}{txseq_list}(object)
}
\value{
An object of class "states", which is a \code{\link{data.table}} with 
the following columns:
\describe{
\item{state_id}{The state ID number.}
\item{state_name}{The state name,}
\item{state_name_long}{A long-form state name.}
}
}
\description{
Create a \code{\link{data.table}} containing information on health states for
use in \code{\link[hesim]{hesim_data}}.
}
\examples{
txseq1 <- txseq(first = "erlotinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "gefitinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab")) 
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2)
states <- create_states(txseqs) 
class(states)
print(states)
}
