% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesim_data.R
\name{create_states}
\alias{create_states}
\title{Create health states table}
\usage{
create_states(start_line = c("first", "second"))
}
\arguments{
\item{start_line}{The line of treatment that a patient is starting.}
}
\value{
An object of class "states", which is a \code{\link{data.table}} with 
the following columns:
\describe{
\item{state_id}{The state ID number.}
\item{state_name}{The state name,}
\item{state_name_long}{A long-form state name.}
}
}
\description{
Create a \code{\link{data.table}} containing information on health states for
use in \code{\link[hesim]{hesim_data}} as a function of whether the patient
is beginning first or second line treatment..
}
\examples{
create_states(start_line = "first")
create_states(start_line = "second")
}
