% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_disease.R
\name{create_trans_mat}
\alias{create_trans_mat}
\title{Create a line-specific transition matrix}
\usage{
create_trans_mat(line = c("1", "2", "2+"))
}
\arguments{
\item{line}{The line of treatment.}
}
\value{
A matrix where rows/columns denote states and entries in the matrix
denote transitions between states.
}
\description{
Create a transition matrix in the same format as the \link[mstate]{mstate} 
conditional on the line of treatment.
}
\examples{
txseq1 <- txseq(first = "erlotinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "gefitinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab")) 
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2)
tmat <- create_trans_mat(line = "2")
class(tmat)
print(tmat)
}
