% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{create_trans_mats}
\alias{create_trans_mats}
\title{Create transition matrcies}
\usage{
create_trans_mats(object)
}
\arguments{
\item{object}{Object used to create transition matrices.}
}
\value{
The form of the returned value depends on \code{object}.
}
\description{
Generic function to create transition matrices.
}
\seealso{
\code{\link{create_trans_mats.txseq_list}}
}
\keyword{internal}
