% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_params.R
\name{sample_params}
\alias{sample_params}
\title{Sample parameters for PSA}
\usage{
sample_params(n, params_mstate_nma = iviNSCLC::params_mstate_nma)
}
\arguments{
\item{n}{The number of random observations of the parameters to draw.}

\item{params_mstate_nma}{A list of \code{\link{params_surv}} objects, 
where each element in the list denotes a survival distribution. Should have
the same variable names as \code{\link{params_mstate_nma}}.}
}
\value{
An object of class "sampled_params", which is a list with the
following components. 
\describe{
\item{mstate_nma}{A \code{\link{params_surv}} object where the number of 
rows in each \code{coef} element is equal to \code{n}.}
}
}
\description{
Sample parameters from their joint probability distribution for a 
probabilistic sensitivity analysis (PSA).
}
\examples{
params <- sample_params(2)
names(params)

# Multi-state NMA
params$mstate_nma$weibull$coefs$shape
params$mstate_nma$weibull$dist
}
