% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatments}
\alias{treatments}
\title{Treatments}
\format{A data.table object with 1 row for each treatment and w columns:
\describe{
  \item{name}{Name of treatment.}
  \item{abb}{Abbreviation for the treatment.}
}}
\usage{
treatments
}
\description{
Dataset of treatments for EGFR positive NSCLC.
}
\keyword{datasets}
